[channelz] More ephemeral logging (#39672)
This change adds a twist on channelz event logging.
Often what we need to capture is some kind of ephemeral stuck state - and for that we need detailed information on an event already completed.

If we capture that for all time, it's too noisy, too much memory, not a good plan.

Instead: add a concept of a commit to log entries in channelz.
All traces are rendered into channelz when the query is made, but uncommitted traces are discarded once their refs go to zero -- that is when some operation is done.
This allows us to track detailed information in the historical trace whilst things are happening, *and* should an error occur - but in the normal path clean up everything once its done and not clog up the system.

Supporting change: make trace collection hierarchical, so that large swaths of traces can be removed when the parent is (that is, to stay in the trace, all nodes up to the top most must become committed).

Whilst I'm there: change from pre-rendered traces to lazily rendered traces - these will usually be smaller (especially for anything complex), and allow us to offset the cost of rendering until we actually need to.

Closes #39672

COPYBARA_INTEGRATE_REVIEW=https://github.com/grpc/grpc/pull/39672 from ctiller:plog 0c248a0684d10f4c148d94c181940f32cda7a0d7
PiperOrigin-RevId: 767650886